cocailon

+++ BackEnd/src/main/java/com/restaurant/backend/Dto/Response/PaymentLinkResponse.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -13,6 +12,6 @@ public class PaymentLinkResponse {
     private String checkoutUrl;
     private String qrCode;
     private Long orderCode;
-    private BigDecimal amount;
+    private Long amount;
 }
